Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -9,7 +9,7 @@ int main(){
     int add = a+b;
     int subtract = a-b;
     int multitply = a*b;
-    float division = a/b;
+    float division = float(a/b);
     if(op=='+'){
         printf("%d",add);
     }
